MB-31869: Don't notify the connection when DCP FlowControl is disabled
As detailed in the MB, DCP shows a relevant throughput degradation if
we disable the DCP FlowControl in memcached.
Root cause of the issue is that we end up calling
Consumer::scheduleNotify() at every incoming DCP Mutation, ~25% of the
total runtime of the mc:worker thread at Consumer as measured by the
profile in my local env.

Note that in the DCP Mutation code-path notifying the connection is
necessary for sending DCP Buffer Ack messages. When FlowControl is
disabled we can just skip the call.

Change-Id: I46754c112fa1e9fd6eec08ec58d1e767b9452742
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/135692
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>