MB-39505: Simplify associating StatCollectors with a bucket/scope/col
Relying on the caller to correctly set a label with a key "bucket",
"scope" or "collection" leaves open the possibility of forgetting to do
so, making a typo, etc.

As the only labels which are currently required are bucket, scope, and
collection, simplify the interface to

 forBucket(bucket)
 forScope(scopeID)
 forCollection(collectionID)

This ensures consistency in both the label keys, and the formatting of
the values - scope and collections will both always be formatted
correctly, with no chance for accidentally providing a name instead,
for example.

By introducing different types {Bucket,Scope,Col}StatCollector (as
thin wrappers around a "real" StatCollector) methods can declare which
labels they require. I.e., for a method

addFooCollectionsStats(ColStatCollector& collector);

The caller _must_ have added bucket, scope, and collection labels to
have created a suitable collector to pass as a parameter.

Change-Id: I69ee2f41828cae92304b5f7cf39f3f6b3a13a741
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137974
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>