[refactor] Rename SERVER_HANDLE_V1 to ServerApi
Back in the days we wanted the core to be able to load thirdparty
engines and needed to provide an API which was binary compatible
so that engines could operate with different versions of the server.
These days we build (and link) everything together so this isn't
a priority anymore. Rename the struct to ServerApi and kill all of
the typedefs to simplify the code.

Change-Id: I1ce9273a5989f0794d6638f6897040ca32193dc1
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/135741
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>