MB-41321: Combine Collection objects 'CommitData' and 'Flush'
The CommitData which is a member of KVStore is now moved into
Collections:VB::Flush. This is required upstream in this sequence of
changes for MB-41321 so the flusher can detect if a flushed key is in a
batch that also drops the key's collection.

Part of this merging also moves some of the persisted data representing
'seqno' to be unsigned types - we never persist a negative seqno so
there is no need to consider negative numbers when doing comparisons of
persisted seqnos.

Change-Id: Id7e1925d5520949fa874c1a384670cbed6f61821
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/136098
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>