MB-38428: Allow compaction and flusher in parallel
We're compacting the current database file into the new file
without holding any locks. The thread performing the compaction
holds the file open so it can't go away while it is traversing
the data in there, and it knows where in the file the "current"
header is located so it also knows where to stop reading data.

Once all of the data up to the "current" header is written it'll
grab the lock and append the mutations which happened after
compaction started.

Change-Id: I40722522d60a0d0065193b19c5940c9892a0b0a1
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/135954
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>