MB-36572: Use CouchKVStoreFileCache for reading
To track files we open for reads against our file descriptor limits we
could do one of a few things:

1) Store the Db/DbHolder in the CouchKVStoreFileCache
2) Reduce the limit of the CouchKVStoreFileCache by a fixed amount
3) Reduce the limit of the CouchKVStoreFileCache by 1 for each read

Solution 1 is problematic; CouchKVStoreFileCache is a map which means
that each file entry needs a unique key. Each read would need to create
some unique key for the cache (bucket name + vBucket isn't enough).
Furthermore, most reads don't actually care about caching a file, they
just want to read some key and go. This means there is not much point
in caching a file anyway.

Solution 2 would be quite simple, but we could easily run into issues
with different workload patterns (i.e. read or write heavy)

Solution 3 would only decrease the limit as we do each read (and the
limit would then be increased as each reader finishes with it's file).
This puts the burden of maintaining the lifetime of the file on the
reader (should they wish to do so) and this works well with existing
tasks such as warmup and backfills that do this. If the file cache
ever gets full, new entries (both reads and writes) will free the
longest not used entry from the cache (under lock to prevent races).
We can decrease the cache size limit by calling resize on the cache.

Change-Id: Ied860c427f02b75cc2c52ec58b8ff8f39a8476bd
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/136174
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>