MB-31869: Make ConnMap::connNotifier const
At http://review.couchbase.org/c/kv_engine/+/101187 we introduce some
new code path that accesses the ConnNotifier from bgthreads.
That leads to a data-race between create-bucket (fethread) and some of
the task executed in bgthreads (eg, PagingVisitor).

The problem is that currently the ConnMap initialization is split in 2
phases: the ctor and the initialize() method. In particular,
ConnMap::connNotifier is contructed in ConnMap::initialize(), which
is executed *after* the EP ExecutorPool and some tasks (eg,
PagingVisitor) are up and running.
When we introduce a call to ConnMap from the PagingVisitor, that races
with the bucket creation on ConnMap::connNotifier.

Given that ConnMap::connNotifier is intrinsically const, then make it
so and construct it in ConnMap ctor. No further synchronization is
required to guard the lifetime of that member.

CV picked that up for http://review.couchbase.org/c/kv_engine/+/101187
on TSAN in the ep_perfsuite "Expiry pager latency" test.

Change-Id: I0d11346d58e4363f726b43f1b983054ec40c7db9
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/136373
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>