MB-41459: Set default magma flushers/compactors correctly
When storage threads configuration is not set, we need to set the
default number of storage threads to magma_max_default_storage_threads
rather than attempting to base the number of storage threads on the
number of writer threads. The default number of writer threads can have
a huge range from a default of 4 to when couchbase is optimized for io
where its set to the number of threads on the machine. Trying to set
storage threads based on number of writer threads probably doesn't make
sense.

When testing magma or for users on small systems using magma, setting
storage threads is the correct way to adjust the number of storage
threads. If not set, the default setting is still small enough that it
should work for almost all cases other than extremely small scenarios
where magma isn't really targeted.

Change-Id: I0c32e084317b46c0d71dd7e189b8458bfbe925e2
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/136436
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>