MB-41510: Fix infinite loop due to HdrHistogram being reset
Currently if a HdrHistogram has its reset() method called by one thread
(T1) and is being read by another (T2) using an HdrHistoram::Iterator.
There is the race that could cause the HdrHistoram::Iterator() to enter
an infinite loop.

This occurs as HdrHistogram's getNextValue() methods use HdrHistogam_c's
hdr_iter_next() which intern uses the following function:
static bool has_next(struct hdr_iter* iter)
{
        return iter->cumulative_count < iter->total_count;
}
Which logically checks that the total count the Iterator was
initialize with is > than the sum of all the counts its read from the
buckets in the histogram. The trouble is, that the reset() method
changes the bucket counts back to 0. Thus, meaning the condition
cumulative_count >= total_count may never be met.

To fix this wrap the std::unique_ptr<struct hdr_histogram> with a
folly::Synchronized<>. This allows us to take an exclusive lock on the
pointer when performing a reset() of the histogram and allows us to
ensure that no iterator is reading from the histogram, as the iterator
takes a non exclusive read lock for the lifetime of the
HdrHistogram::Iterator struct. This also ensures that
HdrHistogram::resize() is thread safe as we now take a write lock.

Change-Id: I8cf5eeb0ac3107ce20e915463be3db245c68c8b9
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137228
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>