MB-41405: Read persistedDeletes from RW store
Currently when we query the KVStore for the persisted deletes stat we
query the RO store which has its own cache and does not track persisted
deletes (as they are tracked in the RW store). This means that querying
persisted deletes returns the first value correctly (as we read it from
disk) but every subsequent query will return the same value. Correct
this by querying persisted deletes from the RW store.

Change-Id: I2524ca0df1b23a428a34bc645dd606ae48ddaefe
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/136523
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>