MB-39745: Notify flusher hpvbs from outside flusher
Currently if there is a high priority request then we set a counter
in the shard and add an element to the VBucket to track this an notify
the flusher. The flusher then runs, checks the shard counter, then adds
the vb to the hpvbs queue. To allow us to remove the shard checks from
the flusher to decouple them, we need to move this notification of a
high priority vb out of the flusher.

Change-Id: Ic238ac020d95dd518c7ef2eb1e2a087dd7e54832
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/136652
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>