Refactor: Use MemoryBackedBuffer in subdoc
Previously we had two variables representing the "current"
state of the document: in_doc and temp_doc inside the
command context.

Every time we changed the underlying document we had to
remember to update temp_doc, and then reset in_doc to
represent this document.

By doing so we would also need two parameters in the call-path
to operate_single_doc; as it would need to change both variables
in order to stay in sync.

This patch put the two variables inside a MemoryBackedBuffer
which allows to start off with a read-only view of the data;
and later reset it to point to a different backing (and keep
the backing with the same lifetime as the view). (initially
subdoc points to the actual document, and we don't need a
new copy if we're not going to modify the document)

Change-Id: I245173a136bf97722e081da3248a6fd4da8643ef
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/136693
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>