MB-41364: Don't lock the bucket for item transfer
Previously we would lock the entire bucket instance and increment
the clients counter to prevent the bucket from being deleted while
a client still had a item reference to the bucket in its send
queue. That logic would cause contention on the bucket lock if
we had multiple clients doing requests on the same bucket trying
to receive documents > 4k (smaller documents would be injected
inline in the stream)

Change the logic to use an std::atomic counter for the number of
items currently in transfer, and wait for it to become 0 when
all of the client references to the bucket disappears.

Change-Id: Ie4a3d19cd7654865f8aeabcb8205bbb892fd46de
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/136785
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>