MB-39292: 3/n Change from JSON to flatbuffers+crc32c
Change from a JSON format to a flatbuffers defined structure.
Reload of the manifest can use the flatbuffers Verify which
checks that the loaded data appears to be a valid 'Manifest'.

A CRC (using crc32c) is further added to the stored data so
that we can detect changes to the data which may not be
detected by Verify (e.g. a string value "dog" changing to "cat").

Detected issues are logged with CRITICAL - the bucket is allowed to
continue warm-up and the in-memory Bucket::Manifest is initialised as
the 'epoch' Manifest, the bucket will then pick-up the current
manifest from ns_server, any Manifest is a successor of epoch.

Change-Id: I319c3e76ef5c07da0680e602f3e342d3d8affa58
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137162
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>