MB-40493: Enforce compression when Value::NoWithUnderlyingDatatype
Currently we enforce passive compression based on the datatype of the
final value (after original item has been possibly uncompressed for
making any modification).
But, IncludeValue::NoWithUnderlyingDatatype may restore the Snappy
datatype before we check if passive compression is required, so we may
miss to re-compress the value that is actually streamed over the
network.

Note that the issue affects the Xattr chunk only. Ie, if we have an
item with Body+Xattr, then IncludeValue::NoWithUnderlyingDatatype will
remove the Body and we will end up with an uncompressed Xattr chunk.
Note also that the max size of Xattrs is 1MB (over the Body size).

Change-Id: I3b8d6f851365fd8095f08792a5cfb111afc15abe
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/136861
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>