MB-41235: Ensure we don't complete a SyncWrite if it don't have a cookie
Currently ActiveDurabilityMonitor::commit() will try and call
VBucket::notifyClientOfSyncWriteComplete() to get it to return
ENGINE_SYNC_WRITE_AMBIGUOUS to the client. However, if the following
steps occur, the Active Durability Monitor may have ActiveSyncWrite
objects in its resolvedQueue that have the cookie pointer set to
nullptr.

1. Replica vbucket has a Prepare sent to it
2. SyncWrite is added to the PDM
3. The collection the prepare is for is removed
4. The replica vbucket is promoted to a active one, turning the PDM into
an ADM. At this point we turn the prepared SyncWrite to an
ActiveSyncWrite, setting its cookie pointer member to nullptr.
5. The ADM receives a majority number of acks for the prepared item
6. ActiveDurabilityMonitor::checkForCommit() runs and adds the value to
the resolvedQueue
7. Then processCompletedSyncWriteQueue() is called and runs though the
resolvedQueue, trying to complete by calling
ActiveDurabilityMonitor::commit() with the ActiveSyncWrite with the
nullptr cookie member.
8. Intern we call notifyClientOfSyncWriteComplete() which throws a gsl
exception as it expects to see a non null cookie.

To fix this we shouldn't try and notify the client as we don't have a
any record of how to contact them. Instead we should return out of this
function.

This patch also adds a test to prove the steps above reproduce this bug
and that this fix works.

Change-Id: Iecea1f1e595c32a4bca865a61072157f5da45b46
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/136925
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>