MB-41713: Remove the database files in each test batch
It is ns_server which is responsible for deleting the files
when a bucket is deleted, and memcached will pick up the
files on disk if someone tries to create a bucket (thats
how the data survives restart of memcached).

There was a problem when running testapp that everyone tried
to create a bucket named "default" in each test batch,
but left the system in an unknown state at the end
of the test batch (which could lead the next test batch
to fail with unknown reasons). This would not be detected by
CV as it runs each test batch in a separate process.

Change-Id: I87eb8284b2b37fcb49783adbe650b0cd8562e45c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137030
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>