MB-41719: Create consumer in ConnMap in test_not_using_backfill_queue
When we call addStream we put the Stream in the ConnStore. As this test
suite runs the full engine, we will on other threads do things such
as flush items. Flushing an item potentially wakes the stream for seqno
acking which requires that the associated Consumer be in the ConnStore.

Fix this individual test by creating the consumer in the ConnMap. This
is potentially a problem in other tests, but this test appears to have
been failing often recently.

Change-Id: I78d5865f49f74e8bbeba9815cbcbfe3dd104a24a
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137054
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>