[testapp]: remove excessive calls to getConnection()
Each call to getConnection() performs a number of roundtrips
to the server (connection establishment, setting connection
properties via hello etc).

A lot of the "helper functions" in the system grabbed a new
connection just to avoid the extra parameter (and didn't
really need a new connection).

This is a preparation for when we'll drop support for the
"default bucket" and access to data without authenticating
(as with that change the "getConnection" gets even more
expensive as it'll also need authentication and select
bucket which is at least two roundtrips to the server)

Change-Id: I843ba7b75f8857f8a7cb368df4d1d5a534c73230
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137115
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>