MB-40285: Allow pruning all elements of the FailoverTable
Allow pruning all elements of the FailoverTable and insert a new element
at the current high seqno.

At rollback we prune all entries in the FailoverTable with a seqno
greater than the new high seqno. It's possible to rollback to a seqno
lower than the first entry in the FailoverTable as the FailoverTable
has a maximum size (by default 25 elements). Currently if we were to
prune all of the FailoverTable entries we would throw an exception. As
this is a valid situation, remove the exception and instead insert a
new element into the FailoverTable should a rollback result in there
being no entries. Once the replica reconnects to the active vBucket it
will pick up the active's copy of the FailoverTable. We insert a single
element into the FailoverTable instead of leaving it empty as much of
the FailoverTable code relies on their being at least 1 entry in the
table.

Change-Id: I51800377a20b57b80bfc127de169c3b3c2d19e82
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137158
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>