Disable concurrent flusher and compaction in couch-kvstore
In MB-41747 it looks like on_disk_prepares is corrupt.

The value gets incremented as part of storing an object to
disk, and decremented as part of compaction when it gets
purged. One of the changes we've done lately is that we
now run the flusher and compaction in parallel so lets start
by disabling that.

Change-Id: Icf39aa4dcfcca29671d8afcd4b42f402369d6050
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137197
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>