MB-39505: Add privilege checking to StatCollector interface
CBStats must limit which scopes/collections stats are returned for
based upon the privileges of the authed user. This was checked through
use of the cookie.

For Prometheus metrics, the MetricServer already requires the authed
user to have the Stats privilege, allowing the collection of _all_
system stats including all scopes and collections. Prometheus
collectors do not have a cookie to test privileges on, and don't need
to check for SimpleStats privilege on a per-scope/collection basis.

Code using StatCollector can no longer rely upon a cookie being
available, so add support for privilege checking through the
StatCollector interface, to allow differing implementations.

Change-Id: I62dc21687078ddbdf0940f41726b86162dcc5cdc
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137978
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>