MB-16181: Restructure Collections::Manifest
The class 'usability' has never been great since the addition of scopes
with collection meta-data split over the collection map and the scopes
map. For example the cmd_stat handler has to do a search for collection
TTL because the data is not in the collection map (now fixed).

This commit updates the object so that there is one 'CollectionEntry'
that holds all collection meta-data, the collection map now just
references that one entry whilst the scopes structure is the owner
of the data.

Change-Id: I547da9020c982804c803c673f3e610c6756a7207
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137499
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>