MB-39745: Scale flushers independently of shards
Currently we create one Flusher object for every KVShard that we create.
This is because historically a Flusher belonged to a KVShard and only
a single flusher would ever attempt to flush against a given
KVShard/KVStore at any given time. For magma, we want to implement
group commit which amortizes fsync costs by grouping together writes
to the same KVStore/WAL. With various recent changes to KVStore, we
have allowed for multiple Flushers to be able to flush in parallel to
the same KVStore object. We will support magma group commit by allowing
multiple Flushers to flush in parallel to a given KVStore. As such, we
must make Flushers scalable independently from KVShards. Each VBucket
will still map by modulo to a given KVShard, and will also map by modulo
to a given Flusher. The intent is that the number of flushers is
increased to some multiple of the number of shards to benefit from
group commit.

Change-Id: I6576443a560685064246a0246a0bd906db609bc2
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137427
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>