MB-41857: Increase cache limit on close of old db during compaction
During compaction we call openDbForRead for the old database file and
then manually close the file. Manually closing the file causes us to
not track the closure of the file against the file cache limit. This
causes us to decrease the file cache limit with no way to later increase
it, causing opens to later fail as the cache is full.

Change-Id: I3b795263d44e62865e537bc75bd11cb84e107759
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137569
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>