Prefix audit stats in Prometheus
"enabled" and "dropped_events" are unambiguous under cbstats, where
there is additional context of being in the audit stat group.

In contrast, there is no such grouping under Prometheus, so it is not
immediately clear what these stats are.

Prefixing these stats with "audit_" for  Prometheus clarifies what they
are.

(cbstats keys unchanged)

Change-Id: I3b5588e9bc085d092f7d24b5e5504e3c8f8c330d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137652
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>