MB-41944: Item::removeUserXattrs() operates on a copy
Item::removeUserXattrs is executed in the ActiveStream path for DCP
connections that set IncludeDeletedUserXattrs::No.

When we make any change to the payload being streamed, we must operate
on a copy of Item::value. Our changes will reflect to other connections
and to the persistence cursor otherwise, as Item::value points to the
shared in-memory blob referenced by items in the CheckpointManager.

Change-Id: I5e8ec8df788b695a3388fdfd95c4db9c79dd0849
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137837
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>