MB-41989: document_pre_expiry() operates on a copy
Expiration removes the UserXattrs (if any) from the item's payload.
Currently we modify the shared blob pointed by the item in the HT
and CM, which is incorrect as changes may reflect on the pre-existing
mutation being persisted, replicated and exposed on the frontend.

Change-Id: I5fd2d3d039a4f0949f6e823592513098b457b09d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137972
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>