MB-39292: 2/n Order set_collections with warm-up
Delay the acceptance of a new manifest from ns_server
whilst warm-up has yet to progress far enough.

This ensures that set_collections follows the loading
of any previously given manifest and KV can assert
that the new manifest is a valid successor.

Without this change KV can accept a manifest and then
proceed to load an older one from disk and fail within
warm-up as we cannot assign the older manifest over the
new one.

Change-Id: Ief2235b875001a77d9a184bb0b598e7fefbc1f5f
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/138166
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>