MB-41718: Fix crash due to vbucket nullptr dereference
Currently in DcpProducer::getHighSeqnoOfCollections() there is the
possibility that we can end up dereferencing a nullptr as we don't check
if getVBucket() returns a valid ptr. To fix this refactor
getHighSeqnoOfCollections() to take a ref to a VBucket object as this
method is only ever called within DcpProducer::streamRequest() which
gains a valid VBucketPtr already. Also move getHighSeqnoOfCollections()
to the protected namespace as its only ever used within DcpProducer.

Change-Id: I74f86662929503d5d81f04296b4ce30cf1870920
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/138055
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>