MB-41699: Prevent ephemeral vb from purging drop of default collection
Differences between ephemeral and persistent vbuckets are too many
for the previous attempt at this functionality (implemented only
in VB::Manifest::queueCollectionSystemEvent) and a mistake in how
unit-tests trigger a full ephemeral 'purge' mean this issue was
missed.

1) Move the code which sets the delete time to -1 to be in
addSystemEventItem which has an ephemeral/persistent implementation.

2) For ephemeral set the "delete or complete time"

3) For persistent we do what was previously done, set the expiry time
which in turn becomes the persisted tomebstone's delete time.

4) Fix the unit-tests so that both of the ephemeral purging tasks
are executed

Change-Id: Ifaca936b941ed1440c49590556aeb4b94542a6ef
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/138383
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>