MB-41728: Use logarithmic buckets for exposing histograms to Prometheus
Percentile based iteration of HdrHistograms provides valuable detail,
but does not lead to consistent bucket bounds and is therefore
unacceptable for Prometheus.

Buckets changing over time leads to a large number of Prometheus time
series, and leads to problems when querying.

Furthermore, percentiles do not necessarily lead to unique bucket
bounds. This leads to a single prometheus scrape finding multiple
results for the same apparent lower bound (as encountered in MB-41728).

Change-Id: Ib6a42478a79500506e64dc9b96e5bcc6407080ca
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/138295
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>