Use cb::couchstore::seek(End) to detect concurrent write
As part of the catch up phase while running concurrent compaction
we want to detect if the flusher wrote data to the database while
the compactor was copying data without holding the lock.

Instead of iterating over _all_ 4k blocks on disk from the current
header until the end of the file to find the last header block it
is more efficient to just try to locate the last header on the
file (which should be at the end of the file).

Improve the logging messages to include if the flusher is being
paused or not while copying the data.

Change-Id: I515a513f6606d1b4bb8d21fa41f69f5e3c41a987
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/138071
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>