MB-41550: Allow SubdocReplaceBodyWithXattr to revive documents
Allow SubdocReplaceBodyWithXattr to operate on deleted documents,
but given that a deleted document can't have a value (only XATTRs)
we also introduce a new flag called ReviveDocument.

The primary motivation for creating a new flag and not just have
SubdocReplaceBodyWithXattr automatically revive a deleted document
is to make sure that it behaves the same way as the other subdoc
operations on deleted documents (in case we ever want to allow
deleted documents to contain value in the future).

Unfortunately we can't do CAS replace when transitioning from a
deleted document to a live document (ep-engine don't support that).
The way the commit phase for transactions work today is currently
that they try to run an Add from the client SDK field (which would
fail if someone had created the document in the mean time). That
is a potential problem that it won't detect if someone creates
and then deletes the document in the time gap from the staging
to the commit. Hopefully we can close the gap by implementing a
cas replace of a document from the Deleted state to Alive.

Change-Id: I17624667d310137e9e04a6acb3f6464789b1829d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/138229
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>