MB-42016: Prevent flusher from missing insert into droppedCollections
A crash occurred because the VB:Manifest placed an item into the
checkpoint and then into a local container (droppedCollections).
This is because the sequence number of the item is needed.
However in-between the two events, the flusher was able to run,
flush the drop collection event and then 'peek' back into the
VB::Manifest - it doesn't find any information about the dropped
collection and throws an exception.

To fix this we need to ensure that the droppedCollection container
is given the sequence number of the event in a different way to
prevent the described bug. This commit uses a std::function callback
that is executed by the checkpoint manager whilst it is locked for
updates - i.e. the checkpoint we are inserting into cannot be flushed
until queueDirty returns - and in this case we will have inserted the
correct sequence number into the VB::Manifest container.

Change-Id: I47349bee51e2f43c39c1282aea98e2237e629535
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/138314
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>