MB-41915: Reduce intermittent failure rate of eviction test
Avoid intermittent failure of
http://review.couchbase.org/c/kv_engine/+/138400

A previous change http://review.couchbase.org/c/kv_engine/+/138400
was insufficient; the test was still seen failing intermittently.

The replica memory usage needs to be low enough to _ensure_ active
vbuckets will be evicted from, but the amount of memory which needs to
be reclaimed includes the active items too.

Change-Id: Iac55d3b353316b58a374f7f1ebe0b9653de71ac0
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/139723
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>